@@ -1,15 +1,24 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
     "./src/**/*.{js,ts,jsx,tsx}",
     "./src/app/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx,css}",
-     "./src/components/Empleados/**/*.{js,ts,jsx,tsx,css}",
+    "./src/components/Empleados/**/*.{js,ts,jsx,tsx,css}",
   ],
   theme: {
-    extend: {
-     
-    },
+
+    // screens: {
+    //   xs: "480px",
+    //   sm: "640px",
+    //   md: "768px",
+    //   lg: "1024px",
+    //   xl: "1280px",
+    //   "2xl": "1536px",
+    //   "3xl": "1920px",
+    //   "4xl": "2560px",
+    // },
+    
   },
   plugins: [],
 };
